refactor: reorganize routines feature into sub-features structure
- Create sub-features: routine-creation, routine-detail, trail-running, workout-management, routine-list
- Move main components from root to appropriate sub-features
- Move all components to sub-feature-specific component directories
- Update all import paths to use absolute paths from root
- Create index files for each sub-feature to export components
- Update main index.ts to export from sub-features
- Fix naming conflicts between types and components (WorkoutSection)
- Resolve all TypeScript compilation errors
- Fix critical linting errors (return types, unused imports)
- Maintain all existing functionality while improving code organization
