+++ apps/web/src/features/routines/features/routine-creation/components/RoutineDetailsForm.tsx
