+++ apps/web/src/features/routines/features/routine-creation/index.ts
@@ -1,0 +1,9 @@
+// Main component
+export { default as RoutineCreation } from './RoutineCreation';
+
+// Components
+export { default as RoutineHeader } from './components/RoutineHeader';
+export { default as RoutineDetailsForm } from './components/RoutineDetailsForm';
+export { default as RoutineModals } from './components/RoutineModals';
+export { default as StrengthWorkoutsSection } from './components/StrengthWorkoutsSection';
+export { default as RunningWorkoutsSection } from './components/RunningWorkoutsSection';
