+++ apps/web/src/features/routines/features/routine-detail/components/RoutineProgress.tsx
