+++ apps/web/src/features/routines/features/routine-detail/components/WeeklySchedule.tsx
