+++ apps/web/src/features/routines/features/routine-detail/index.ts
@@ -1,0 +1,11 @@
+// Main component
+export { default as RoutineDetail } from './RoutineDetail';
+
+// Components
+export { default as RoutineDetailHeader } from './components/RoutineDetailHeader';
+export { default as RoutineInfo } from './components/RoutineInfo';
+export { default as RoutineOverviewCards } from './components/RoutineOverviewCards';
+export { default as RoutineProgress } from './components/RoutineProgress';
+export { default as WeeklySchedule } from './components/WeeklySchedule';
+export { default as WorkoutDaysList } from './components/WorkoutDaysList';
+export { default as ExerciseList } from './components/ExerciseList';
