+++ apps/web/src/features/routines/features/routine-list/index.ts
@@ -1,0 +1,8 @@
+// Main component
+export { default as Routines } from './Routines';
+
+// Components
+export { default as RoutineCard } from './components/RoutineCard';
+export { default as RoutinesList } from './components/RoutinesList';
+export { default as ActiveRoutineCard } from './components/ActiveRoutineCard';
+export { default as WorkoutCard } from './components/WorkoutCard';
