+++ apps/web/src/features/routines/features/trail-running/components/RepeatIntervalsForm.tsx
@@ -1,6 +1,9 @@
 'use client';
 
+import { Plus, Trash2 } from 'lucide-react';
+
 import { Button } from '@/components/ui/button';
+import { Card, CardContent, CardHeader, CardTitle } from '@/components/ui/card';
 import { Input } from '@/components/ui/input';
 import { Label } from '@/components/ui/label';
 import {
@@ -13,7 +16,10 @@ import {
 import { Switch } from '@/components/ui/switch';
 
 import IntensityTargetConfiguration from './IntensityTargetConfiguration';
-import { TrailRunningInterval, IntensityTarget } from '../types';
+import {
+  TrailRunningInterval,
+  IntensityTarget,
+} from '@/features/routines/types';
 
 interface RepeatIntervalsFormProps {
   repeatIntervals: Partial<TrailRunningInterval>[];
