+++ apps/web/src/features/routines/features/trail-running/index.ts
@@ -1,0 +1,13 @@
+// Main component
+export { default as TrailRunningWorkout } from './TrailRunningWorkout';
+
+// Components
+export { default as TrailRunningHeader } from './components/TrailRunningHeader';
+export { default as WorkoutStats } from './components/WorkoutStats';
+export { default as WorkoutOverview } from './components/WorkoutOverview';
+export { default as IntensityTargetConfiguration } from './components/IntensityTargetConfiguration';
+export { default as SectionTypeSelector } from './components/SectionTypeSelector';
+export { default as RepeatIntervalsForm } from './components/RepeatIntervalsForm';
+export { default as SectionsList } from './components/SectionsList';
+export { default as SectionForm } from './components/SectionForm';
+export { default as SectionFormCard } from './components/SectionFormCard';
