+++ apps/web/src/features/routines/features/workout-management/components/WorkoutDetails.tsx
@@ -1,6 +1,9 @@
 'use client';
 
+import { Card, CardContent } from '@/components/ui/card';
+import { Input } from '@/components/ui/input';
 import { Label } from '@/components/ui/label';
+import { Textarea } from '@/components/ui/textarea';
 import {
   Select,
   SelectContent,
@@ -8,8 +11,6 @@ import {
   SelectTrigger,
   SelectValue,
 } from '@/components/ui/select';
-import { Textarea } from '@/components/ui/textarea';
-import { Input } from '@/components/ui/input';
 
 interface WorkoutDetailsProps {
   objective: string;
