+++ apps/web/src/features/routines/features/workout-management/index.ts
@@ -1,0 +1,5 @@
+// Components
+export { default as WorkoutHeader } from './components/WorkoutHeader';
+export { default as WorkoutDetails } from './components/WorkoutDetails';
+export { default as WorkoutSection } from './components/WorkoutSection';
+export { default as ExerciseManagement } from './components/ExerciseManagement';
